[Fix] change cannotDeleteTagWhenExistedTask

+++ src/main/java/com/nhnacademy/mini_dooray/taskapi/service/tag/TagServiceImpl.java
@@ -4,21 +4,25 @@ import com.nhnacademy.mini_dooray.taskapi.dto.tag.TagIndexRequestDto;
 import com.nhnacademy.mini_dooray.taskapi.dto.tag.TagRequestDto;
 import com.nhnacademy.mini_dooray.taskapi.entity.Project;
 import com.nhnacademy.mini_dooray.taskapi.entity.Tag;
+import com.nhnacademy.mini_dooray.taskapi.entity.TaskTag;
 import com.nhnacademy.mini_dooray.taskapi.exception.project.NotFoundProjectException;
 import com.nhnacademy.mini_dooray.taskapi.exception.tag.NotFoundTagException;
 import com.nhnacademy.mini_dooray.taskapi.repository.ProjectRepository;
 import com.nhnacademy.mini_dooray.taskapi.repository.TagRepository;
+import com.nhnacademy.mini_dooray.taskapi.repository.TaskTagRepository;
 import java.util.List;
+import java.util.Set;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.web.bind.annotation.RequestBody;
 
 @Service
 @RequiredArgsConstructor
 public class TagServiceImpl implements TagService {
     private final TagRepository tagRepository;
     private final ProjectRepository projectRepository;
-
+    private final TaskTagRepository taskTagRepository;
     @Override
     public List<TagIndexRequestDto> getTagListByProjectId(Long projectId) {
         List<Tag> tags = tagRepository.findAllByProject_ProjectId(projectId);
@@ -36,14 +40,23 @@ public class TagServiceImpl implements TagService {
     }
 
     @Override
-    public TagRequestDto getTagByTagId(Long tagId) {
-        return new TagRequestDto(
-                tagRepository.findById(tagId).orElseThrow(() -> new NotFoundTagException("태그를 찾을 수 없습니다"))
-                        .getTagName());
+    public List<TagRequestDto> getTagsByProejctIdAndTaskId(Long projectId, Long taskId){
+        List<Tag> tags = tagRepository.findAllByProject_ProjectId(projectId);
+        List<TaskTag> taskTags = taskTagRepository.findAllByPk_TaskId(taskId);
+
+        Set<Long> taskTagIds = taskTags.stream()
+                .map(taskTag -> taskTag.getTag().getTagId())
+                .collect(Collectors.toSet());
+
+        return tags.stream()
+                .filter(tag -> taskTagIds.contains(tag.getTagId()))
+                .map(tag -> new TagRequestDto(tag.getTagName()))
+                .collect(Collectors.toList());
     }
 
+
     @Override
-    public Tag saveTag(Long projectId, TagRequestDto tagRequest) {
+    public Tag saveTag(Long projectId, @RequestBody TagRequestDto tagRequest) {
         Project project = projectRepository.findById(projectId)
                 .orElseThrow(() -> new NotFoundProjectException(projectId + "project를 찾을 수 없습니다"));
 
@@ -63,6 +76,9 @@ public class TagServiceImpl implements TagService {
 
     @Override
     public void deleteTag(Long projectId, Long tagId) {
+        if (taskTagRepository.existsByPk_TagId(tagId)) {
+            throw new RuntimeException("태그를 사용하는 태스크가 있습니다");
+        }
         if (checkProjectId(projectId, tagId)) {
             tagRepository.deleteById(tagId);
         } else {
